Do not require getSystemInfo() for ISO15693 tags.
ISO15693 tags that do not respond to the
getSystemInfo() command should still be reported
to the upper stack, as getSystemInfo is not mandatory.

This also adds some code to deal with specifc ST
tag configurations, that require the prot ext bit
to be set when executing some commands.

+++ src/nfa/int/nfa_rw_int.h
@@ -188,14 +188,14 @@ typedef struct
     UINT8               *p_block_data;
 } tNFA_RW_OP_PARAMS_T3T_WRITE;
 
-/* NFA_RW_OP_I93_INVENTORY params */
+/* NFA_RW_OP_I93_XXX params */
 typedef struct
 {
     BOOLEAN             uid_present;
     UINT8               uid[I93_UID_BYTE_LEN];
     UINT8               afi;
     UINT8               dsfid;
-    UINT8               first_block_number;
+    UINT16              first_block_number;
     UINT16              number_blocks;
     UINT8              *p_data;
 } tNFA_RW_OP_PARAMS_I93_CMD;
